unit test for errors page controllers and errors page accessibility.
