+++ package.json
@@ -15,10 +15,10 @@
     "lint": "eslint . --ext .ts",
     "build": "yarn webpack --config webpack.config.js",
     "build:prod": "NODE_ENV=production yarn webpack --mode production --config webpack.config.js",
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "test:routes": "TS_NODE_TRANSPILE_ONLY=true NODE_ENV=production LOG_LEVEL=OFF mocha --opts mocha.opts $(find src/test \\( -name '*.ts' \\) -path '*/routes/*') --timeout 10000 --reporter-options reportFilename=routes,inlineAssets=true,reportTitle=CaT-buyer-frontend-routes",
-    "test:a11y": "TS_NODE_TRANSPILE_ONLY=true NODE_ENV=production LOG_LEVEL=OFF mocha --exit --opts mocha.opts src/test/a11y/a11y.ts --timeout 60000 --reporter-options reportFilename=a11y,inlineAssets=true,reportTitle=CaT-buyer-frontend-a11y",
+    "test": "SET NODE_ENV=development && mocha --opts mocha.opts src/test/**/*.test.ts --timeout 10000 --reporter-options reportFilename=routes,inlineAssets=true,reportTitle=CaT-buyer-frontend-unit-test-all",
+    "test:coverage": "SET NODE_ENV==development nyc --reporter=text npm test",
+    "test:routes": "SET NODE_ENV=development && mocha --opts mocha.opts src/test/routes/**/*.test.ts --timeout 10000 --reporter-options reportFilename=routes,inlineAssets=true,reportTitle=CaT-buyer-frontend-routes",
+    "test:a11y": "SET NODE_ENV=development && mocha --exit --opts mocha.opts src/test/a11y/a11y.test.ts --timeout 60000 --reporter-options reportFilename=a11y,inlineAssets=true,reportTitle=CaT-buyer-frontend-a11y",
     "sonar-scan": "sonar-scanner",
     "test:smoke": "echo 'TODO add smoke tests' && exit 1",
     "test:functional": "echo 'TODO add functional tests' && exit 1",
@@ -83,7 +83,6 @@
     "@types/chai-as-promised": "^7.1.1",
     "@types/copy-webpack-plugin": "^6.0.0",
     "@types/express-session": "^1.17.0",
-    "@types/jest": "^26.0.13",
     "@types/mini-css-extract-plugin": "^0.9.1",
     "@types/mocha": "^9.0.0",
     "@types/pa11y": "^5.3.3",
@@ -115,7 +114,6 @@
     "gulp-replace": "^1.1.3",
     "gulp-sass": "^5.0.0",
     "html-webpack-plugin": "^4.3.0",
-    "jest": "^27.3.1",
     "mini-css-extract-plugin": "^0.11.2",
     "mocha": "^6.2.2",
     "mocha-junit-reporter": "^1.23.3",
@@ -135,7 +133,6 @@
     "sonar-scanner": "^3.1.0",
     "style-loader": "^1.1.3",
     "supertest": "^4.0.2",
-    "ts-jest": "^25.2.1",
     "ts-loader": "^6.2.1",
     "tslint": "5.20.1",
     "typescript": "^4.5.2",
