+++ src/main/errors/model/errorConstants.ts
@@ -1,0 +1,5 @@
+export enum ERROR_PATHS {
+    //@ All of the Path related Request for authorization
+    ROUTE_404 = '/404',
+    ROUTE_401 = '/401'
+}
