--- src/main/errors/model/routeconstants.ts
@@ -1,5 +1,0 @@
-export enum ROUTECONST  {
-    //@ All of the Path related Request for authorization
-    Route_404 = '/404',
-    Route_401 = '/401'
-}
