+++ src/main/errors/path.ts
@@ -1,6 +1,6 @@
 
 import { ERROR_CONTROLLER } from './controller/index'
-import { ROUTECONST } from './model/routeconstants'
+import { ERROR_PATHS } from './model/errorConstants'
 import { Application } from 'express';
 import { NO_AUTH } from '../common/middlewares/oauthservice/openroutecheck'
 
@@ -9,9 +9,9 @@ export default function (app: Application): void {
 
 
   //@GET '/404'
-  app.get(ROUTECONST.Route_404, ERROR_CONTROLLER.Error_404)
+  app.get(ERROR_PATHS.ROUTE_404, ERROR_CONTROLLER.Error_404)
 
   //@GET '/401'
-  app.get(ROUTECONST.Route_401, NO_AUTH, ERROR_CONTROLLER.Error_401)
+  app.get(ERROR_PATHS.ROUTE_401, NO_AUTH, ERROR_CONTROLLER.Error_401)
 
 }
